rtmp: support Extended Metadata Frames (#4006) (#4014)
