+++ internal/protocols/rtmp/message/reader_test.go
@@ -277,6 +277,25 @@ var readWriterCases = []struct {
 			0x31, 0x01, 0x02, 0x03,
 		},
 	},
+	{
+		"extended metadata",
+		&ExtendedMetadata{
+			ChunkStreamID:   0x6,
+			DTS:             0,
+			MessageStreamID: 0x1000000,
+			FourCC:          0x68766331,
+			Payload:         []interface{}{"colorInfo", amf0.Object{amf0.ObjectEntry{Key: "colorConfig", Value: amf0.Object{}}}},
+		},
+		[]byte{
+			0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x26, 0x09,
+			0x01, 0x00, 0x00, 0x00, 0x84, 0x68, 0x76, 0x63,
+			0x31, 0x02, 0x00, 0x09, 0x63, 0x6f, 0x6c, 0x6f,
+			0x72, 0x49, 0x6e, 0x66, 0x6f, 0x03, 0x00, 0x0b,
+			0x63, 0x6f, 0x6c, 0x6f, 0x72, 0x43, 0x6f, 0x6e,
+			0x66, 0x69, 0x67, 0x03, 0x00, 0x00, 0x09, 0x00,
+			0x00, 0x09,
+		},
+	},
 }
 
 func TestReader(t *testing.T) {
