+++ internal/protocols/rtmp/message/testdata/fuzz/FuzzReader/105635bd94dfe048
@@ -1,0 +1,2 @@
+go test fuzz v1
+[]byte("0000\x00\x00\b\t0000\x94hvc1000")
